Move StyleTags names and insertRule helpers to separate files
