+++ src/utils/once.js
@@ -1,6 +1,6 @@
 // @flow
 // Helper to call a given function, only once
-export default (cb: () => any) => {
+export default (cb: () => void): (() => void) => {
   let called = false
 
   return () => {
